Reorganize form examples into Form Layouts, Form Controls, and Form Validation.
