--- src-docs/src/views/field_controls/field_number.js
@@ -1,34 +1,0 @@
-import React, {
-  Component,
-} from 'react';
-
-import {
-  EuiFieldNumber,
-} from '../../../../src/components';
-
-export default class extends Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      value: '',
-    };
-  }
-
-  onChange = e => {
-    const sanitizedValue = parseInt(e.target.value, 10);
-    this.setState({
-      value: isNaN(sanitizedValue) ? '' : sanitizedValue,
-    });
-  };
-
-  render() {
-    return (
-      <EuiFieldNumber
-        placeholder="Placeholder text"
-        value={this.state.value}
-        onChange={this.onChange}
-      />
-    );
-  }
-}
