+++ src-docs/src/views/form_controls/select.js
@@ -1,0 +1,73 @@
+import React, {
+  Component,
+  Fragment,
+} from 'react';
+
+import {
+  EuiSelect,
+  EuiSpacer,
+} from '../../../../src/components';
+
+import makeId from '../../../../src/components/form/form_row/make_id';
+
+export default class extends Component {
+  constructor(props) {
+    super(props);
+
+    this.options = [
+      { value: 'option_one', text: 'Option one' },
+      { value: 'option_two', text: 'Option two' },
+      { value: 'option_three', text: 'Option three' },
+    ];
+
+    this.state = {
+      value: this.options[1].value,
+    };
+  }
+
+  onChange = e => {
+    this.setState({
+      value: e.target.value,
+    });
+  };
+
+  render() {
+    return (
+      <Fragment>
+        <EuiSelect
+          options={this.options}
+          value={this.state.value}
+          onChange={this.onChange}
+        />
+
+        <EuiSpacer size="m" />
+
+        <EuiSelect
+          options={this.options}
+          value={this.state.value}
+          onChange={this.onChange}
+          disabled
+        />
+
+        <EuiSpacer size="m" />
+
+        <EuiSelect
+          options={this.options}
+          value={this.state.value}
+          onChange={this.onChange}
+          isLoading
+        />
+
+        <EuiSpacer size="m" />
+
+        <EuiSelect
+          options={this.options}
+          value={this.state.value}
+          onChange={this.onChange}
+          isLoading
+          disabled
+        />
+      </Fragment>
+    );
+  }
+}
