+++ src-docs/src/views/form_layouts/form_layouts_example.js
@@ -51,7 +51,7 @@ const inlinePopoverHtml = renderToHtml(InlinePopover);
 export const FormLayoutsExample = {
   title: 'Form layouts',
   sections: [{
-    title: 'Form rows',
+    title: 'Form and form rows',
     source: [{
       type: GuideSectionTypes.JS,
       code: formRowsSource,
@@ -61,8 +61,8 @@ export const FormLayoutsExample = {
     }],
     text: (
       <p>
-        Use the <EuiCode>FormRow</EuiCode> component to easily associate form components with
-        labels, help text, and error text.
+        Use the <EuiCode>EuiFormRow</EuiCode> component to easily associate form components with
+        labels, help text, and error text. Use the <EuiCode>EuiForm</EuiCode> component to group <EuiCode>EuiFormRow</EuiCode>s.
       </p>
     ),
     props: {
